refactor: move cli commands into their own module
