+++ iroh/Cargo.toml
@@ -35,6 +35,7 @@ console = { version = "0.15.5", optional = true }
 dirs-next = { version = "2.0.0", optional = true }
 hex = { version = "0.4.3", optional = true }
 indicatif = { version = "0.17", features = ["tokio"], optional = true }
+iroh-io = { path = "../iroh-io", optional = true }
 multibase = { version = "0.9.1", optional = true }
 num_cpus = { version = "1.15.0", optional = true }
 postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"], optional = true }
@@ -45,7 +46,7 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"], optional = tr
 
 [features]
 default = ["cli", "metrics"]
-cli = ["clap", "config", "console", "dirs-next", "hex", "indicatif", "iroh-bytes/io", "multibase", "num_cpus", "postcard", "quic-rpc/quinn-transport", "range-collections", "tempfile", "tokio/rt-multi-thread", "tracing-subscriber"]
+cli = ["clap", "config", "console", "dirs-next", "hex", "indicatif", "iroh-io", "multibase", "num_cpus", "postcard", "quic-rpc/quinn-transport", "range-collections", "tempfile", "tokio/rt-multi-thread", "tracing-subscriber"]
 metrics = ["iroh-metrics"]
 test = []
