+++ iroh/src/commands/doctor.rs
